update
I noticed recently a registration error on phpBB2x.com. While the sessions bug does exist, the failure is caused by a mod that is yet to be determined and ill fix it soon. However in debugging it I noticed some other bugs and irritating deficiencies in phpbb2x so I deconstructed it and fixed a few.
1. malformed sessions ids that result in sporadically being logged out. FIXED
2. firefoxes incessant malformed Lax cookies warning FIXED
3. admin link redirecting to root/index instead of root/admin/index FIXED finally after how many years?
A. Cookie Compliance - For GDPR and (CCPA) compliance I added a cookie consent banner that can be enabled/disabled in the ACP?General section.
B. Stop Forum Spam - I added a registration block that uses Stop Forum Spam db of spam ip's, email addresses and usernames, also enable-able in the ACP

If you already have a phpBB2x installation, just remove the root/config.php from the new files and overwrite the current files (*there are also a few minor template fixes ) and run this in phpmyadmin

    INSERT INTO phpbb_config (config_name, config_value) VALUES ('enable_cookie_consent', '1');
    INSERT INTO phpbb_config (config_name, config_value) VALUES ('enable_sfs_check', '1');